MOB-1341: move hooks to ExploreV2 directory
