+++ src/components/Explore/ExploreV2/hooks/useLocationSearch.ts
